requirements_full.txt
#To train the model, run Grad-CAM, YOLO training and optionally use labelImg GUI
# ---------------- TensorFlow & Keras ----------------
tensorflow==2.10.0
tensorflow-estimator==2.10.0
keras==2.10.0
Keras-Preprocessing==1.1.2
opt_einsum==3.4.0

# ---------------- PyTorch & YOLO ----------------
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118
ultralytics==8.3.177
ultralytics-thop==2.0.15

# ---------------- Image & Visualization ----------------
opencv-python==4.11.0.86
Pillow==11.3.0
matplotlib==3.10.5

# ---------------- Data Handling ----------------
numpy==1.26.4
pandas==2.3.1
scikit-learn==1.7.1
scipy==1.15.3

# ---------------- Grad-CAM & Explanation ----------------
# Grad-CAM uses TensorFlow and matplotlib mainly
# Already included above

# ---------------- Utilities ----------------
tqdm==4.67.1
pyyaml==6.0.2
requests==2.32.4

# ---------------- Optional GUI & Labeling ----------------
PyQt5==5.15.11
labelImg @ git+https://github.com/HumanSignal/labelImg.git@b33f965b6d14c14f1e46b247f1bf346e03f2e950

# ---------------- TensorBoard ----------------
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
